drawberu-processor/requirements.txt
flask==3.1.1
flask-cors==5.0.1
gunicorn==23.0.0
opencv-python-headless==4.11.0.86
numpy==2.2.6
Pillow==11.2.1
scikit-learn==1.7.0
scikit-image==0.25.2
